I fixed some react warnings and added uid to handle unic ids to the attribute props
