+++ package.json
@@ -15,7 +15,8 @@
     "react-redux": "^7.1.3",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
-    "styled-components": "^5.1.1"
+    "styled-components": "^5.1.1",
+    "uid": "^1.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
